Laporan Update

+++ database/migrations/2025_06_14_094155_create_registrasis_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('nomer_induk_nasabah')->nullable();
             $table->string('password')->nullable();
             $table->date('tanggal');
-            $table->bigInteger('saldo')->nullable();
+            $table->decimal('saldo', 15, 2)->nullable();
             $table->string('foto')->nullable();
             $table->timestamps();
         });
